Adding more work on the oopool. BugzID: 423

+++ src/main/java/com/laureninnovations/util/IOUtil.java
@@ -10,17 +10,19 @@ public class IOUtil {
      * @throws IOException If there is an error deleting a file or directory.
      */
     public void delete(File file) throws IOException {
-        if (file.isDirectory()) {
-            File[] files = file.listFiles();
-            if (files != null) {
-                for (File c : files) {
-                    delete(c);
+        if (file.exists()) {
+            if (file.isDirectory()) {
+                File[] files = file.listFiles();
+                if (files != null) {
+                    for (File c : files) {
+                        delete(c);
+                    }
                 }
             }
-        }
 
-        if (!file.delete()) {
-            throw new FileNotFoundException("Failed to delete file: " + file);
+            if (!file.delete()) {
+                throw new FileNotFoundException("Failed to delete file: " + file);
+            }
         }
     }
